requirements.txt
Flask==2.1.0
gunicorn==20.1.0
flask-cors==3.1.1
pdfplumber==0.5.28
pytesseract==0.3.9
Pillow==9.0.1
